+++ src/components/SmishingTemplates/DeleteTemplateModal.vue
@@ -10,7 +10,7 @@
     @changeStatus="closeModal"
   >
     <template #app-dialog-body>
-      {{ selectedEmailTemplate && selectedEmailTemplate.name }} will be deleted.
+      {{ getBodyText }}
     </template>
     <template #app-dialog-footer>
       <app-dialog-footer
@@ -41,6 +41,16 @@ export default {
     },
     selectedEmailTemplate: {
       type: Object
+    },
+    templateCount: {
+      type: Number,
+      default: 0
+    },
+    isMultiple: {
+      type: Boolean
+    },
+    multipleDeletePayload: {
+      type: Object
     }
   },
   data() {
@@ -48,20 +58,38 @@ export default {
       isActionButtonDisabled: false
     }
   },
+  computed: {
+    getBodyText() {
+      if (this.isMultiple) {
+        return `${this.templateCount} smishing templates will be deleted.`
+      }
+      return `${this.selectedEmailTemplate && this.selectedEmailTemplate.name} will be deleted.`
+    }
+  },
   methods: {
     closeModal() {
-      this.$emit('handleCloseModal')
+      this.$emit('on-close')
     },
     handleDelete() {
-      this.isActionButtonDisabled = true
-      SmishingService.deleteTextMessageTemplate(this.selectedEmailTemplate.resourceId)
-        .then(() => {
-          this.$emit('handleSuccessDeleteAction', this.selectedEmailTemplate)
-          this.closeModal()
-        })
-        .finally(() => {
-          this.isActionButtonDisabled = false
-        })
+      if (this.isMultiple) {
+        this.isActionButtonDisabled = true
+        SmishingService.bulkDeleteTextMessageTemplates(this.multipleDeletePayload)
+          .then(() => {
+            this.$emit('on-success-multiple')
+          })
+          .finally(() => {
+            this.isActionButtonDisabled = false
+          })
+      } else {
+        this.isActionButtonDisabled = true
+        SmishingService.deleteTextMessageTemplate(this.selectedEmailTemplate.resourceId)
+          .then(() => {
+            this.$emit('on-success', this.selectedEmailTemplate)
+          })
+          .finally(() => {
+            this.isActionButtonDisabled = false
+          })
+      }
     }
   }
 }
